fix: provide semantic highlight with list of ranges (#99)
To provide semantic highlights, we have to resolve a list of ranges
instead of a single range. To do that we can simply take the min start
and max end of the range list. VSCode is doing the exact same thing.

rzls also change the endpoint name of the handler to match this new
behavior(note the plural on "ranges").